+++ frontend/src/components/task-detail/TaskHeader.test.tsx
@@ -67,7 +67,6 @@ function renderHeader(
       setEditMaxCostUsd={noop}
       editPriority={0}
       setEditPriority={noop}
-      runs={[]}
       taskSaving={false}
       taskSaveError=""
       onStartEdit={noop}